requirements.txt
tqdm
pillow
opencv-python
gradio
ftfy
regex
torch
torchvision